--- src/components/post/write/StyledQuill.css
@@ -1,27 +1,0 @@
-.ql-toolbar {
-  background: #fff;
-  width: 950px;
-  height: 52px;
-
-  border-radius: 10px 10px 0px 0px;
-  border-bottom: 1px solid #d9d9d9;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.ql-container {
-  background: #fff;
-  flex-grow: 1;
-  width: 950px;
-  padding-bottom: 0px;
-  border-radius: 0px 0px 10px 10px;
-
-  display: flex;
-  flex-direction: column;
-}
-
-.ql-editor {
-  font-family: 'Pretendard-Regular', serif;
-}
