+++ src/hooks/useLoginUserId.ts
@@ -1,7 +1,4 @@
 import { useEffect, useMemo, useState } from 'react';
-import { useAtom } from 'jotai';
-import { userAtom } from 'src/globalState/jotai';
-import supabase from 'src/lib/supabaseClient';
 
 const useLoginUserId = () => {
   const [loginUserId, setLoginUserId] = useState<string>('');
