+++ src/pages/Register.tsx
@@ -17,4 +17,5 @@ const Register = () => {
     </>
   );
 };
+
 export default Register;
